requirements.txt
tensorflow==2.17.0
paho-mqtt==2.1.0
numpy==1.26.4
flask==3.0.3 
requests==2.32.3